requirements/base.txt
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
flake8==3.5.0
Markdown==2.6.11
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
selenium==3.14.0
urllib3==1.23
